fix/추천 1 상세 referenceTitle 필드추가 및 추천, 탐색 api 요청 시 lastValue 타입 string 변경

+++ src/main/java/com/anipick/backend/explore/dto/ExploreItemDto.java
@@ -9,6 +9,6 @@ public class ExploreItemDto {
 	private Long    id;
 	private String  title;
 	private String  coverImageUrl;
-	private Integer averageScore;
+	private Double averageScore;
 	private Long popularId;
 }
